refactor: types/rezervoTypes => types/rezervo
