+++ components/modals/Agenda/Agenda.tsx
@@ -3,7 +3,7 @@ import AgendaClassItem, { AgendaClass } from "./AgendaClassItem";
 import React from "react";
 import { classConfigRecurrentId } from "../../../lib/iBooking";
 import { SitClass } from "../../../types/integration/sit";
-import { ClassConfig } from "../../../types/rezervoTypes";
+import { ClassConfig } from "../../../types/rezervo";
 import { WEEKDAY_NUMBER_TO_NAME } from "../../../utils/timeUtils";
 
 export default function Agenda({
