+++ components/modals/ClassInfo/ClassInfoModal.tsx
@@ -1,7 +1,7 @@
 import { Modal } from "@mui/material";
 import ClassInfo from "./ClassInfo";
 import { sitClassRecurrentId } from "../../../lib/iBooking";
-import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex } from "../../../types/rezervoTypes";
+import { AllConfigsIndex, ClassPopularity, ClassPopularityIndex } from "../../../types/rezervo";
 import React, { Dispatch, SetStateAction } from "react";
 import { SitClass } from "../../../types/integration/sit";
