+++ components/schedule/WeekSchedule.tsx
@@ -1,7 +1,7 @@
 import { Box, Stack } from "@mui/material";
 import DaySchedule from "./DaySchedule";
 import React, { useEffect } from "react";
-import { AllConfigsIndex, ClassPopularityIndex } from "../../types/rezervoTypes";
+import { AllConfigsIndex, ClassPopularityIndex } from "../../types/rezervo";
 import { SitClass, SitWeekSchedule } from "../../types/integration/sit";
 import { useRouter } from "next/router";
