+++ components/schedule/class/ClassPopularityMeter.tsx
@@ -1,6 +1,6 @@
 import SpeedRoundedIcon from "@mui/icons-material/SpeedRounded";
 import QuestionMarkRoundedIcon from "@mui/icons-material/QuestionMarkRounded";
-import { ClassPopularity, StatusColors } from "../../../types/rezervoTypes";
+import { ClassPopularity, StatusColors } from "../../../types/rezervo";
 import { SitClass } from "../../../types/integration/sit";
 import { isClassInThePast } from "../../../lib/iBooking";
 import { stringifyClassPopularity, determineClassPopularity } from "../../../lib/popularity";
