+++ components/schedule/class/ClassUserAvatar.tsx
@@ -2,7 +2,7 @@ import { Avatar, Badge } from "@mui/material";
 import { hexColorHash } from "../../../utils/colorUtils";
 import ErrorRoundedIcon from "@mui/icons-material/ErrorRounded";
 import RippleBadge from "../../utils/RippleBadge";
-import { StatusColors } from "../../../types/rezervoTypes";
+import { StatusColors } from "../../../types/rezervo";
 import React from "react";
 
 export default function ClassUserAvatar({
