+++ hooks/useUserConfig.ts
@@ -1,5 +1,5 @@
 import useSWR from "swr";
-import { ConfigPayload, UserConfig } from "../types/rezervoTypes";
+import { ConfigPayload, UserConfig } from "../types/rezervo";
 import { fetcher } from "../utils/fetchUtils";
 import { useUser } from "@auth0/nextjs-auth0/client";
 import useSWRMutation from "swr/mutation";
