+++ hooks/useUserSessions.ts
@@ -1,5 +1,5 @@
 import useSWR from "swr";
-import { UserSessionsIndex } from "../types/rezervoTypes";
+import { UserSessionsIndex } from "../types/rezervo";
 import { fetcher } from "../utils/fetchUtils";
 import { useUser } from "@auth0/nextjs-auth0/client";
