+++ lib/iBooking.ts
@@ -1,6 +1,6 @@
 import { GROUP_BOOKING_URL, TIME_ZONE } from "../config/config";
 import { SitClass, SitWeekSchedule } from "../types/integration/sit";
-import { ClassConfig } from "../types/rezervoTypes";
+import { ClassConfig } from "../types/rezervo";
 import { weekdayNameToNumber } from "../utils/timeUtils";
 import { DateTime } from "luxon";
 import { calculateMondayOffset } from "./integration/common";
